chore: move duration to its own module
